src/main/webapp/app/entities/usersms/restaurant-user/update/restaurant-user-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder, Validators } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { IRestaurantUser, RestaurantUser } from '../restaurant-user.model';
import { RestaurantUserService } from '../service/restaurant-user.service';
import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

@Component({
  selector: 'jhi-restaurant-user-update',
  templateUrl: './restaurant-user-update.component.html',
})
export class RestaurantUserUpdateComponent implements OnInit {
  isSaving = false;

  usersSharedCollection: IUser[] = [];

  editForm = this.fb.group({
    id: [null, [Validators.required, Validators.minLength(5)]],
    internalUser: [],
  });

  constructor(
    protected restaurantUserService: RestaurantUserService,
    protected userService: UserService,
    protected activatedRoute: ActivatedRoute,
    protected fb: FormBuilder
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ restaurantUser }) => {
      this.updateForm(restaurantUser);

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const restaurantUser = this.createFromForm();
    if (restaurantUser.id !== undefined) {
      this.subscribeToSaveResponse(this.restaurantUserService.update(restaurantUser));
    } else {
      this.subscribeToSaveResponse(this.restaurantUserService.create(restaurantUser));
    }
  }

  trackUserById(index: number, item: IUser): string {
    return item.id!;
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IRestaurantUser>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(restaurantUser: IRestaurantUser): void {
    this.editForm.patchValue({
      id: restaurantUser.id,
      internalUser: restaurantUser.internalUser,
    });

    this.usersSharedCollection = this.userService.addUserToCollectionIfMissing(this.usersSharedCollection, restaurantUser.internalUser);
  }

  protected loadRelationshipsOptions(): void {
    this.userService
      .query()
      .pipe(map((res: HttpResponse<IUser[]>) => res.body ?? []))
      .pipe(map((users: IUser[]) => this.userService.addUserToCollectionIfMissing(users, this.editForm.get('internalUser')!.value)))
      .subscribe((users: IUser[]) => (this.usersSharedCollection = users));
  }

  protected createFromForm(): IRestaurantUser {
    return {
      ...new RestaurantUser(),
      id: this.editForm.get(['id'])!.value,
      internalUser: this.editForm.get(['internalUser'])!.value,
    };
  }
}
